apps/chatbotcsv/requirements.txt
absl-py==1.4.0
aioboto3==10.4.0
aiobotocore==2.4.2
aiofiles==22.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
aiosqlite==0.19.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argilla==1.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
audio-recorder-streamlit==0.0.8
autopep8==2.0.2
azure-cognitiveservices-speech==1.25.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.0
black==23.3.0
bleach==6.0.0
blinker==1.6.2
boto3==1.24.59
botocore==1.27.59
bs4==0.0.1
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.17
cleo==2.0.1
click==8.1.3
clickhouse-connect==0.5.18
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.7
crashtest==0.4.1
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
deeplake==3.3.0
defusedxml==0.7.1
Deprecated==1.2.13
# detectron2==0.6
dill==0.3.6
discord.py==2.2.2
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckdb==0.7.1
duckduckgo-search==2.8.5
dulwich==0.21.3
effdet==0.3.0
elevenlabs==0.2.15
elevenlabslib==0.4.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.7.3
Faker==18.4.0
fastapi==0.92.0
fastjsonschema==2.16.3
filelock==3.10.7
flatbuffers==23.3.3
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.16.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-speech==2.17.3
google-cloud-texttospeech==2.14.1
google-search-results==2.4.2
googleapis-common-protos==1.59.0
gradio_client==0.2.0
gradio_tools==0.0.7
greenlet==2.0.1
grpcio==1.51.3
grpcio-status==1.51.3
gTTS==2.3.1
h11==0.14.0
hnswlib==0.7.0
html2text==2020.1.16
html5lib==1.1
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
humanfriendly==10.0
humbug==0.3.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
installer==0.7.0
iopath==0.1.9
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
keyring==23.13.1
kiwisolver==1.4.4
kor==0.8.0
langchain==0.0.173
layoutparser==0.3.4
llama-index==0.5.5
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
orjson==3.8.10
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.11.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
pkginfo==1.9.6
platformdirs==2.6.2
playsound==1.2.2
playwright==1.32.1
ply==3.11
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
portalocker==2.7.0
posthog==2.4.2
pox==0.3.2
ppft==1.7.6.6
praw==7.7.0
prawcore==2.3.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyairtable==1.4.0
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pydub==0.25.1
pyee==9.0.4
pyflakes==3.0.1
PyGithub==1.58.1
Pygments==2.15.1
PyJWT==2.6.0
Pympler==1.0.1
PyNaCl==1.5.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.7.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytube==12.1.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz==2.15.0
readability-lxml==0.8.1
redis==4.5.3
regex==2023.3.23
replicate==0.8.1
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4
spotipy==2.22.1
SQLAlchemy==2.0.14
stack-data==0.6.2
starlette==0.25.0
streamlit==1.22.0
streamlit-chat==0.0.2.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.3.3
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
trove-classifiers==2023.3.9
twilio==7.17.0
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
unstructured==0.5.8
unstructured-inference==0.2.14
update-checker==0.18.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.20.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.21.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
whisper==1.1.10
wikipedia==1.4.0
Wikipedia-API==0.5.8
wrapt==1.14.1
xattr==0.10.1
XlsxWriter==3.0.9
y-py==0.5.9
yacs==0.1.8
yarl==1.9.2
youtube-transcript-api==0.5.0
ypy-websocket==0.8.2
zipp==3.15.0
zstandard==0.20.0
